Corretto tipo armor

+++ Archer.h
@@ -13,7 +13,7 @@ public:
     Archer(const sf::Texture &archerTexture, const sf::Texture &bowTexture, const sf::Texture &arrowTexture,
            const sf::Vector2i &spawnTile, const sf::Vector2i &tileSize, const sf::Vector2i &rectSkin,
            sf::Vector2f damageHit, const std::vector<std::vector<Node>> &nodeMap, bool animated = true, float hp = 15,
-           float speed = 200.f, unsigned int points = 5, unsigned int coins = 1, int armor = 0, bool bubble = false);
+           float speed = 200.f, unsigned int points = 5, unsigned int coins = 1, float armor = 0, bool bubble = false);
 
     void hit(GameCharacter &target, std::vector<std::unique_ptr<Enemy>> &targetEnemies) override;
 };
