+++ Mike.cpp
@@ -12,7 +12,7 @@ bool Mike::isKillStreak(GameCharacter &character) {
 
 Mike::Mike(const sf::Texture &mikeTexture, const sf::Texture &handgunTexture, const sf::Texture &handgunBulletTexture,
            const sf::Vector2i &spawnTile, const TextureManager &guiTexManager, const sf::Vector2i &tileSize,
-           const sf::Vector2i &rectSkin, float startRoundCountdownSeconds, float hp, float speed, int armor)
+           const sf::Vector2i &rectSkin, float startRoundCountdownSeconds, float hp, float speed, float armor)
         : GameCharacter(mikeTexture, hp,
                         speed,
                         0,
