+++ Warrior.h
@@ -24,7 +24,8 @@ public:
     Warrior(const sf::Texture &warriorTexture, const sf::Texture &shieldTexture, const sf::Vector2i &spawnTile,
             const sf::Vector2i &tileSize, const sf::Vector2i &rectSkin, float defense, sf::Vector2f damageHit,
             const std::vector<std::vector<Node>> &nodeMap, float hitProbability, bool animated = true, float hp = 20,
-            float speed = 230.f, unsigned int points = 10, unsigned int coins = 2, int armor = 5, bool bubble = false);
+            float speed = 230.f, unsigned int points = 10, unsigned int coins = 2, float armor = 5,
+            bool bubble = false);
 
     void receiveDamage(float damagePoints) override;
